made possible to load minitig file from previous computation

+++ src/build_index.hpp
@@ -253,6 +253,7 @@ void do_coloring(string& color_load_file, string& color_dump_file, string& fof, 
 		cout<<"File of file problem"<<endl;
 	}
 	color_number = file_names.size();
+	
 	if (not color_load_file.empty()) //query
 	{
 		ifstream nb_eq_f(nb_eq_class_file);
@@ -290,25 +291,38 @@ kmer_Set_Light* load_rle_index(uint k, string& color_load_file, string& color_du
 // build index from new file
 void build_index(uint k, uint m1,uint m2,uint m3, uint c, uint bit, string& color_load_file, string& color_dump_file, string& fof, bool record_counts, bool record_reads, uint64_t& color_number, kmer_Set_Light& ksl, uint nb_threads, bool exact, string& output, bool do_query_on_disk, bool quantize, bool do_log)
 {
-	cout << "Minitig coloring..."<< endl;
 	high_resolution_clock::time_point t1 = high_resolution_clock::now();
-	// apply minitig merge (-> MMM) with rule regarding colors or counts
-	if (record_counts)
+	bool delete_minitig_file(false);
+	if (not exists_test(output +"/_blmonocolor.fa"))
 	{
-		if (quantize)
+		
+		cout << "Minitig coloring..."<< endl;
+		// apply minitig merge (-> MMM) with rule regarding colors or counts
+		if (record_counts)
 		{
-			ksl.construct_index_fof(fof, output, 2, 0);
+			if (quantize)
+			{
+				ksl.construct_index_fof(fof, output, 2, 0);
+			}
+			else
+			{
+				if (do_log)
+					ksl.construct_index_fof(fof, output, 3, 0);
+				else
+					ksl.construct_index_fof(fof, output, 1, 0);
+			}
 		}
 		else
 		{
-			if (do_log)
-				ksl.construct_index_fof(fof, output, 3, 0);
-			else
-				ksl.construct_index_fof(fof, output, 1, 0);
+			ksl.construct_index_fof(fof, output, 0, 0);
 		}
+	} 
+	else 
+	{
+		cout << "Warning , _blmonocolor.fa file was found in output dir, I will use it and I won't delete it" << endl;
+		ksl.construct_index(output + "/_blmonocolor.fa",output);
+		delete_minitig_file = false;
 	}
-	else
-		ksl.construct_index_fof(fof, output, 0, 0);
 	vector<unsigned char*>compr_minitig_color;
 	vector<unsigned> compr_minitig_color_size;
 	long eq_class_nb(0);
@@ -322,8 +336,11 @@ void build_index(uint k, uint m1,uint m2,uint m3, uint c, uint bit, string& colo
 	cout << "Dump index..."<< endl;
 	ksl.dump_disk(output + "/reindeer_index.gz");
 	high_resolution_clock::time_point t13 = high_resolution_clock::now();
-	string cmd("rm -f " + output +"/_blmonocolor.fa");
-	int sysRet(system(cmd.c_str()));
+	if (delete_minitig_file)
+	{
+		string cmd("rm -f " + output +"/_blmonocolor.fa");
+		int sysRet(system(cmd.c_str()));
+	}
 	duration<double> time_span13 = duration_cast<duration<double>>(t13 - t2);
 	cout<<"Index written on disk: "<< time_span13.count() << " seconds."<<endl;
